feat: create standalone TooltipTrigger and TooltipContent components
