+++ packages/react/src/tooltip/context.js
@@ -1,0 +1,10 @@
+import { createContext } from 'react';
+
+const TooltipContext = createContext();
+
+const TooltipProvider = TooltipContext.Provider;
+
+export {
+  TooltipContext,
+  TooltipProvider,
+};
