+++ packages/react/src/tooltip/styles.js
@@ -1,7 +1,13 @@
 import { useColorMode } from '../color-mode';
 import { useColorStyle } from '../color-style';
 
-const useTooltipStyle = props => {
+const useTooltipTriggerStyle = () => {
+  return {
+    display: 'inline-flex',
+  };
+};
+
+const useTooltipContentStyle = () => {
   const [colorMode] = useColorMode();
   const [colorStyle] = useColorStyle({ colorMode });
   const baseStyle = {
@@ -32,5 +38,6 @@ const useTooltipStyle = props => {
 };
 
 export {
-  useTooltipStyle,
+  useTooltipTriggerStyle,
+  useTooltipContentStyle,
 };
